Changes in text, cloud and alignment of skateboarder
I made the links in the clouds change color so they call more attention and make it more intuitive for the users to click on, since in the usability test I noticed a few users didn't know where to click. I also added a text cloud with more work information. I changed the color of the h1 so users don't confuse it with a link and I aligned the skateboarder to the ground road.